pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "naukri-job-automation"
version = "1.0.0"
description = "Automated job application system for Naukri platform with intelligent filtering and human behavior simulation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Sagaya Raj", email = "ssagayaraj37@gmail.com"}
]
keywords = ["naukri", "automation", "job-search", "job-application"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.1",
    "apscheduler>=3.10.4",
    "nltk>=3.8.1",
    "python-dotenv>=1.0.0",
    "flask>=2.3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
]
deploy = [
    "gunicorn>=20.1.0",
]

[project.scripts]
app = "app:app"

[project.urls]
HomePage = "https://github.com/SAGAYARAJ0/naukri-job-automation"
Repository = "https://github.com/SAGAYARAJ0/naukri-job-automation.git"
Documentation = "https://github.com/SAGAYARAJ0/naukri-job-automation"

[tool.setuptools]
packages = ["core", "modules", "utils"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = [".", "tests"]
python_files = "test_*.py"
